Create Basic Operations In Stack
<h1>Push: </h1>
Adds an item to the stack. If the stack is full, then it is said to be an Overflow condition.

<h2>Algorithm for push:</h2>
<pre>
begin
 if stack is full
    return
 endif
else  
 increment top
 stack[top] assign value
end else
end procedure
</pre>


<h1>Pop: </h1>
Removes an item from the stack. The items are popped in the reversed order in which they are pushed. If the stack is empty, then it is said to be an Underflow condition.

<h2>Algorithm for Pop: </h2>
<pre>
begin
 if stack is empty
    return
 endif
else
 store value of stack[top]
 decrement top
 return value
end else
end procedure
</pre>